requirements.txt
Django==5.2.8
djangorestframework==3.15.2
django-environ==0.11.2
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.8

